Move flash helpers into helper file
