+++ include/flash.php
@@ -1,0 +1,20 @@
+<?php
+
+if (session_status() === PHP_SESSION_NONE) {
+    session_start();
+}
+
+function set_flash(string $type, array $messages): void
+{
+    $_SESSION['flash'] = ['type' => $type, 'messages' => $messages];
+}
+
+function get_flash(): ?array
+{
+    if (empty($_SESSION['flash'])) {
+        return null;
+    }
+    $f = $_SESSION['flash'];
+    unset($_SESSION['flash']);
+    return $f;
+}
